requirements.txt
flask
line-bot-sdk
python-dotenv

openai
